ComputeTangent에 MikkTSpace 알고리즘 적용 중

+++ Samples/RenderingApp/src/Application.cpp
@@ -68,7 +68,7 @@ BOOL Application::OnInit(HWND hWnd, int width, int height)
 	size_t MeshCount = pModelAsset->GetNumMeshes();
 	for (size_t i = 0; i < MeshCount; ++i)
 	{
-		pModelAsset->GetMeshDesc(i, MeshDesc);
+		pModelAsset->GetMeshDesc((uint32_t)i, MeshDesc);
 	}
 
 	pModelAsset->Release();
